PR nits and comments
