+++ src/Apps/Ingress/EntryPoints/index.ts
@@ -1,0 +1,6 @@
+/* 
+    These component exports are manually parsed in the build process to generate direct import entry points.
+    To allow regex parsing, the exports must use the following syntax:
+    export { default as <page_name> } from './<path_to_component>';
+*/
+export { default as IngressApp } from './IngressApp/IngressApp';
