+++ src/Apps/Ingress/Models/Constants.ts
@@ -1,11 +1,3 @@
-export enum PageNames {
-    Home = 'Home',
-    TemplatesForm = 'TemplatesForm',
-    SourceMappingForm = 'SourceMappingForm',
-    EventHandlerForm = 'EventHandlersForm',
-    SubscriptionsForm = 'SubscriptionsForm'
-}
-
 export const INGRESS_TRANSLATION_PATH = 'Ingress';
 export const INGRESS_TRANSLATION_COMMON_PATH = `${INGRESS_TRANSLATION_PATH}.Common`;
 export const INGRESS_STORYBOOK_APPNAME = 'Ingress';
