+++ src/Apps/index.ts
@@ -1,0 +1,1 @@
+export * from './Ingress';
