+++ src/index.ts
@@ -1,4 +1,5 @@
 export * from './Adapters';
+export * from './Apps';
 export * from './Models/Classes';
 export * from './Models/Constants';
 export * from './Models/Context';
